Fix Budget categories ID list https://localhost:7020/Budgets/Create

+++ TMD-WalletMaster.Core/Repositories/CategoryRepository.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
 using TMD_WalletMaster.Core.Data;
 using TMD_WalletMaster.Core.Models;
 using TMD_WalletMaster.Core.Repositories.Interfaces;
-using Microsoft.EntityFrameworkCore;
 
 namespace TMD_WalletMaster.Core.Repositories
 {
@@ -41,8 +41,11 @@ namespace TMD_WalletMaster.Core.Repositories
         public async Task DeleteCategoryAsync(int id)
         {
             var category = await _context.Categories.FindAsync(id);
-            _context.Categories.Remove(category);
-            await _context.SaveChangesAsync();
+            if (category != null)
+            {
+                _context.Categories.Remove(category);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
